% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inference.R
\name{anova.loglik}
\alias{anova.loglik}
\title{ANOVA Method for loglik Objects (Likelihood Ratio Test)}
\usage{
\method{anova}{loglik}(object, ..., test = "LRT")
}
\arguments{
\item{object}{First fitted model (of class \code{loglik}).}

\item{...}{Additional fitted models.}

\item{test}{Character. Must be "LRT" or "Chisq" (both perform LRT).}
}
\value{
Object of class \code{lrt} containing test results.
}
\description{
S3 method for \code{anova()} that performs likelihood ratio tests
on nested models fitted with \code{\link{loglik}}.
}
\details{
This is a wrapper around \code{\link{lrt}} that provides familiar
\code{anova()} syntax while explicitly using likelihood ratio tests.
}
\examples{
\dontrun{
anova(fit1, fit2, test = "LRT")
}

}
