% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{bread.loglik}
\alias{bread.loglik}
\title{Compute Bread Matrix for Sandwich Estimator}
\usage{
bread.loglik(object)
}
\arguments{
\item{object}{Object of class "loglik".}
}
\value{
Bread matrix.
}
\description{
Returns the "bread" component: negative of the Hessian.
}
