% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{bread}
\alias{bread}
\title{Compute Bread Matrix for Sandwich Estimator}
\usage{
bread(object, ...)
}
\arguments{
\item{object}{Object of class "loglik".}

\item{...}{Additional arguments (currently unused).}
}
\value{
Bread matrix.
}
\description{
Returns the "bread" component: negative of the Hessian.
}
