% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstatLogLik-package.R
\docType{package}
\name{bstatLogLik-package}
\alias{bstatLogLik}
\alias{bstatLogLik-package}
\title{bstatLogLik: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
