% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstatLogLik-package.R
\docType{package}
\name{bstatLogLik}
\alias{bstatLogLik-package}
\alias{bstatLogLik}
\title{bstatLogLik: A package for computing likelihood statistics}
\description{
The bstatLogLik package provides functions for computing
log-likelihood and related statistics for Bayesian models.
}
\author{
\strong{Maintainer}: Bert van der Werf \email{bert.vanderwerf@auckland.ac.nz} (\href{https://orcid.org/0000-0003-3072-5937}{ORCID})

}
\keyword{internal}
