% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inference.R
\name{compare_models}
\alias{compare_models}
\title{Compare Multiple Models with Information Criteria}
\usage{
compare_models(..., lrt = TRUE)
}
\arguments{
\item{...}{Fitted models of class \code{loglik}.}

\item{lrt}{Logical. Perform likelihood ratio tests if models are nested (default TRUE).}
}
\value{
Data frame with model comparison statistics.
}
\description{
Compares models using AIC, BIC, and optionally likelihood ratio tests.
}
\examples{
\dontrun{
compare_models(fit1, fit2, fit3)
}

}
