% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{compute_gradient}
\alias{compute_gradient}
\title{Compute Jacobian (Gradient) Matrix for Log-Likelihood}
\usage{
compute_gradient(
  deriv_output,
  parameter_info,
  cond_vec,
  npars,
  parnames_selected,
  control
)
}
\arguments{
\item{deriv_output}{List. Output from Deriv::Deriv containing derivative components.}

\item{parameter_info}{Object of class \code{parameter_info}.}

\item{cond_vec}{Logical vector. Which parameters are being estimated.}

\item{npars}{Integer. Total number of parameter groups.}

\item{parnames_selected}{Character vector. Names of parameters being estimated.}

\item{control}{List from loglik_control().}
}
\value{
Numeric matrix of gradients (n_obs x n_params).
}
\description{
Assembles the gradient matrix from automatic differentiation output,
accounting for parameter design matrices and conditional estimation.
}
