% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_derivatives.R
\name{compute_hessian}
\alias{compute_hessian}
\title{Compute Hessian (Second Derivative) Matrix for Log-Likelihood}
\usage{
compute_hessian(
  deriv_output,
  gradient,
  parameter_info,
  cond_vec,
  npars,
  parnames_selected,
  control
)
}
\arguments{
\item{deriv_output}{List. Output from Deriv::Deriv.}

\item{gradient}{Matrix. Computed gradient/Jacobian.}

\item{parameter_info}{Object of class \code{parameter_info}.}

\item{cond_vec}{Logical vector. Which parameters are being estimated.}

\item{npars}{Integer. Total number of parameter groups.}

\item{parnames_selected}{Character vector. Names of parameters being estimated.}

\item{control}{List from loglik_control().}
}
\value{
Symmetric Hessian matrix (n_params x n_params).
}
\description{
Assembles the Hessian from automatic differentiation or outer product of scores,
with optional Levenberg-Marquardt damping.
}
