% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{control_settings}
\alias{control_settings}
\title{Extract Control Settings}
\usage{
control_settings(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Object of class \code{loglik_control}.
}
\description{
Extract Control Settings
}
