% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{converged}
\alias{converged}
\title{Check Convergence Status}
\usage{
converged(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Logical, TRUE if optimization converged.
}
\description{
Check Convergence Status
}
