% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{df.model.loglik}
\alias{df.model.loglik}
\title{Extract Model Degrees of Freedom}
\usage{
df.model.loglik(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Integer, number of estimated parameters.
}
\description{
Returns the number of estimated parameters.
}
