% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{df.residual.loglik}
\alias{df.residual.loglik}
\title{Extract Residual Degrees of Freedom}
\usage{
\method{df.residual}{loglik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Integer, residual degrees of freedom (n - p).
}
\description{
Extract Residual Degrees of Freedom
}
