% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{estfun}
\alias{estfun}
\title{Extract Score Contributions (estfun) for Sandwich Estimator}
\usage{
estfun(object)
}
\arguments{
\item{object}{Object of class "loglik".}
}
\value{
Matrix of score contributions.
}
\description{
Returns the matrix of score contributions (one row per observation).
}
