% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_function.R
\name{evaluate_function}
\alias{evaluate_function}
\title{Evaluate a Model Function Dynamically}
\usage{
evaluate_function(fun, vars = NULL, parameter_info, data = NULL)
}
\arguments{
\item{fun}{The function to evaluate (e.g. model or cdf).}

\item{vars}{Optional named character vector of variable mapping ("model_var" = "data_column").}

\item{parameter_info}{Metadata created by \code{create_parameter_info()}, containing parameter values, design matrices, links, formulas.}

\item{data}{Data frame containing evaluation data.}
}
\value{
Vector or list of results for the supplied data.
}
\description{
Validates inputs, builds dynamic environments, applies link functions, differentiates if required, and evaluates for a given data frame.
}
