% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{fitted.loglik}
\alias{fitted.loglik}
\title{Extract Fitted Values}
\usage{
\method{fitted}{loglik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Numeric vector of fitted values.
}
\description{
Extract Fitted Values
}
