% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{formula.loglik}
\alias{formula.loglik}
\title{Extract Model Formula(s)}
\usage{
\method{formula}{loglik}(x, ...)
}
\arguments{
\item{x}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
List of formulas for each parameter.
}
\description{
Extract Model Formula(s)
}
