% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_gamma.R
\name{gamma_distribution}
\alias{gamma_distribution}
\title{Gamma Distribution Functions for Likelihood Estimation}
\description{
Probability density function (PDF), cumulative distribution function (CDF),
and utility functions for the two-parameter Gamma distribution with
shape (alpha) and rate (beta) parameterization.
}
\details{
These functions are optimized for use with \code{\link{loglik}} and include
pre-computed derivative rules stored in \code{\link{drule}}.
}
\seealso{
Other distribution functions: 
\code{\link{lognormal_distribution}},
\code{\link{normal_distribution}},
\code{\link{weibull_distribution}}
}
\concept{distribution functions}
