% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_gamma.R
\name{gamma_hazard}
\alias{gamma_hazard}
\title{Gamma Hazard Function}
\usage{
gamma_hazard(x, alpha, beta)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{alpha}{Numeric. Shape parameter (must be positive).}

\item{beta}{Numeric. Rate parameter (must be positive).}
}
\value{
Numeric vector of hazard rates.
}
\description{
Computes the hazard (failure rate) function of the Gamma distribution.
}
\details{
The hazard is \eqn{h(x) = f(x) / (1 - F(x))}.
}
\examples{
gamma_hazard(x = 0:10, alpha = 2, beta = 1)

}
