% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_gamma.R
\name{gamma_mean}
\alias{gamma_mean}
\title{Gamma Mean}
\usage{
gamma_mean(alpha, beta)
}
\arguments{
\item{alpha}{Numeric. Shape parameter (must be positive).}

\item{beta}{Numeric. Rate parameter (must be positive).}
}
\value{
Numeric scalar, the mean of the distribution.
}
\description{
Computes the expected value (mean) of the Gamma distribution.
}
\details{
The mean is:
\deqn{E[X] = \frac{\alpha}{\beta}}
}
\examples{
gamma_mean(alpha = 2, beta = 1)

}
