% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_gamma.R
\name{gamma_variance}
\alias{gamma_variance}
\title{Gamma Variance}
\usage{
gamma_variance(alpha, beta)
}
\arguments{
\item{alpha}{Numeric. Shape parameter (must be positive).}

\item{beta}{Numeric. Rate parameter (must be positive).}
}
\value{
Numeric scalar, the variance of the distribution.
}
\description{
Computes the variance of the Gamma distribution.
}
\details{
The variance is:
\deqn{\text{Var}(X) = \frac{\alpha}{\beta^2}}
}
\examples{
gamma_variance(alpha = 2, beta = 1)

}
