% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{hessian}
\alias{hessian}
\title{Extract Hessian Matrix}
\usage{
hessian(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Numeric matrix, Hessian at the optimum.
}
\description{
Returns the second derivative matrix at the optimum.
Useful for checking curvature and numerical stability of fit.
}
