% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_info_methods.R
\name{links}
\alias{links}
\title{Extract link functions}
\usage{
links(object, ...)
}
\arguments{
\item{object}{A \code{parameter_info} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A named list where each element is the link function
for the corresponding parameter.
}
\description{
Retrieves the link functions applied to each parameter.
}
\details{
Link functions describe transformations applied to linear predictors
to obtain parameter values. Common links include "identity", "log", "logit".
}
\examples{
info <- create_parameter_info(
  parameter_values = c(k = 2, lambda = 10, c = 0),
  link_functions = c("log", "log", "identity")
)

links(info)

}
