% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{logLik.loglik}
\alias{logLik.loglik}
\title{Extract Log-Likelihood from loglik Object}
\usage{
\method{logLik}{loglik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Object of class \code{logLik} with attributes for \code{df} and \code{nobs}.
}
\description{
Extract Log-Likelihood from loglik Object
}
