% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_cdf}
\alias{lognormal_cdf}
\title{Log-Normal Cumulative Distribution Function}
\usage{
lognormal_cdf(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of quantiles (must be positive).}

\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Computes the CDF of the Log-Normal distribution.
}
\details{
The CDF is given by:
\deqn{F(x; \mu, \sigma) = \Phi\left(\frac{\log x - \mu}{\sigma}\right)}

where \eqn{\Phi} is the standard normal CDF.
}
\examples{
lognormal_cdf(x = 1:10, mu = 0, sigma = 1)

}
