% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_distribution}
\alias{lognormal_distribution}
\title{Log-Normal Distribution Functions for Likelihood Estimation}
\description{
Probability density function (PDF), cumulative distribution function (CDF),
and utility functions for the two-parameter Log-Normal distribution with
meanlog (mu) and sdlog (sigma) parameterization.
}
\details{
These functions are optimized for use with \code{\link{loglik}} and include
pre-computed derivative rules stored in \code{\link{drule}}.
}
\seealso{
Other distribution functions: 
\code{\link{gamma_distribution}},
\code{\link{normal_distribution}},
\code{\link{weibull_distribution}}
}
\concept{distribution functions}
