% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_hazard}
\alias{lognormal_hazard}
\title{Log-Normal Hazard Function}
\usage{
lognormal_hazard(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric vector of hazard rates.
}
\description{
Computes the hazard (failure rate) function of the Log-Normal distribution.
}
\details{
The hazard is \eqn{h(x) = f(x) / (1 - F(x))}.
}
\examples{
lognormal_hazard(x = 1:10, mu = 0, sigma = 1)

}
