% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_mean}
\alias{lognormal_mean}
\title{Log-Normal Mean}
\usage{
lognormal_mean(mu, sigma)
}
\arguments{
\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric scalar, the mean of the distribution.
}
\description{
Computes the expected value (mean) of the Log-Normal distribution.
}
\details{
The mean is:
\deqn{E[X] = \exp\left(\mu + \frac{\sigma^2}{2}\right)}
}
\examples{
lognormal_mean(mu = 0, sigma = 1)

}
