% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_median}
\alias{lognormal_median}
\title{Log-Normal Median}
\usage{
lognormal_median(mu)
}
\arguments{
\item{mu}{Numeric. Mean of log-transformed data.}
}
\value{
Numeric scalar, the median of the distribution.
}
\description{
Computes the median of the Log-Normal distribution.
}
\details{
The median is:
\deqn{\text{Median} = e^\mu}
}
\examples{
lognormal_median(mu = 0)

}
