% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_mode}
\alias{lognormal_mode}
\title{Log-Normal Mode}
\usage{
lognormal_mode(mu, sigma)
}
\arguments{
\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric scalar, the mode of the distribution.
}
\description{
Computes the mode of the Log-Normal distribution.
}
\details{
The mode is:
\deqn{\text{Mode} = \exp(\mu - \sigma^2)}
}
\examples{
lognormal_mode(mu = 0, sigma = 1)

}
