% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_pdf}
\alias{lognormal_pdf}
\title{Log-Normal Probability Density Function}
\usage{
lognormal_pdf(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of quantiles (must be positive).}

\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric vector of density values.
}
\description{
Computes the PDF of the Log-Normal distribution.
}
\details{
The PDF is given by:
\deqn{f(x; \mu, \sigma) = \frac{1}{x\sigma\sqrt{2\pi}} \exp\left(-\frac{(\log x - \mu)^2}{2\sigma^2}\right)}

for \eqn{x > 0}.
}
\examples{
lognormal_pdf(x = 1:10, mu = 0, sigma = 1)

}
