% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_quantile}
\alias{lognormal_quantile}
\title{Log-Normal Quantile Function}
\usage{
lognormal_quantile(p, mu, sigma)
}
\arguments{
\item{p}{Numeric vector of probabilities.}

\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric vector of quantiles.
}
\description{
Computes the quantile (inverse CDF) of the Log-Normal distribution.
}
\examples{
lognormal_quantile(p = c(0.025, 0.5, 0.975), mu = 0, sigma = 1)

}
