% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_lognormal.R
\name{lognormal_variance}
\alias{lognormal_variance}
\title{Log-Normal Variance}
\usage{
lognormal_variance(mu, sigma)
}
\arguments{
\item{mu}{Numeric. Mean of log-transformed data.}

\item{sigma}{Numeric. Standard deviation of log-transformed data (must be positive).}
}
\value{
Numeric scalar, the variance of the distribution.
}
\description{
Computes the variance of the Log-Normal distribution.
}
\details{
The variance is:
\deqn{\text{Var}(X) = \left(e^{\sigma^2} - 1\right) e^{2\mu + \sigma^2}}
}
\examples{
lognormal_variance(mu = 0, sigma = 1)

}
