% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inference.R
\name{lrt}
\alias{lrt}
\title{Likelihood Ratio Test for Nested Models}
\usage{
lrt(object, ...)
}
\arguments{
\item{object}{First fitted model (of class \code{loglik}).}

\item{...}{Additional fitted models or arguments.}

\item{test}{Character. Always uses "LRT" for likelihood ratio test.}
}
\value{
Data frame with test results including log-likelihoods, df, test statistic, and p-value.
}
\description{
Performs likelihood ratio test comparing two or more nested models fitted
with \code{\link{loglik}}. Models must be nested (one is a special case of another).
}
\details{
The test statistic is:
\deqn{LR = -2(\ell_0 - \ell_1)}
where \eqn{\ell_0} is the log-likelihood of the restricted (simpler) model
and \eqn{\ell_1} is the log-likelihood of the unrestricted (more complex) model.
Under the null hypothesis (that the restrictions are true), LR follows a
chi-squared distribution with degrees of freedom equal to the difference
in number of parameters.

Models are automatically ordered from simplest (fewest parameters) to most complex.
Each model is compared to the previous (simpler) one.

\strong{Important assumptions:}
\itemize{
\item{Models must be nested (parameters in simpler model are subset of complex model)}
\item{Models must be fit to the same data}
\item{Models must have converged successfully}
\item{Asymptotic chi-squared approximation must be valid (large sample, interior of parameter space)}
}
}
\examples{
\dontrun{
# Fit nested models
pars1 <- create_parameter_info(c(k = 2, lambda = 5), formulas = list(k = ~1, lambda = ~1))
pars2 <- create_parameter_info(c(k = 2, lambda = 5), formulas = list(k = ~group, lambda = ~1))

fit1 <- loglik(weibull_pdf, "time", pars1, data)
fit2 <- loglik(weibull_pdf, "time", pars2, data)

# Likelihood ratio test
lrt(fit1, fit2)
}

}
