% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{meat.loglik}
\alias{meat.loglik}
\title{Compute Meat Matrix for Sandwich Estimator}
\usage{
meat.loglik(object)
}
\arguments{
\item{object}{Object of class "loglik".}
}
\value{
Meat matrix.
}
\description{
Calculates the "meat" component: sum of outer products of scores.
}
