% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{meat}
\alias{meat}
\title{Compute Meat Matrix for Sandwich Estimator}
\usage{
meat(object, ...)
}
\arguments{
\item{object}{Object of class "loglik".}

\item{...}{Additional arguments (currently unused).}
}
\value{
Meat matrix.
}
\description{
Calculates the "meat" component: sum of outer products of scores.
}
