% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{model.frame.loglik}
\alias{model.frame.loglik}
\title{Extract Model Data}
\usage{
\method{model.frame}{loglik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Data frame used in model fitting.
}
\description{
Returns the data frame used for model fitting.
}
