% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{nobs.loglik}
\alias{nobs.loglik}
\title{Extract Number of Observations}
\usage{
\method{nobs}{loglik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Integer, number of observations used in fitting.
}
\description{
Extract Number of Observations
}
