% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_cdf}
\alias{normal_cdf}
\title{Normal Cumulative Distribution Function}
\usage{
normal_cdf(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{mu}{Numeric. Mean parameter.}

\item{sigma}{Numeric. Standard deviation parameter (must be positive).}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Computes the CDF of the Normal distribution.
}
\details{
The CDF is given by:
\deqn{F(x; \mu, \sigma) = \Phi\left(\frac{x - \mu}{\sigma}\right)}

where \eqn{\Phi} is the standard normal CDF.
}
\examples{
normal_cdf(x = -3:3, mu = 0, sigma = 1)

}
