% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_distribution}
\alias{normal_distribution}
\title{Normal Distribution Functions for Likelihood Estimation}
\description{
Probability density function (PDF), cumulative distribution function (CDF),
and utility functions for the two-parameter Normal (Gaussian) distribution
with mean (mu) and standard deviation (sigma) parameterization.
}
\details{
These functions are optimized for use with \code{\link{loglik}} and include
pre-computed derivative rules stored in \code{\link{drule}}.
}
\seealso{
Other distribution functions: 
\code{\link{gamma_distribution}},
\code{\link{lognormal_distribution}},
\code{\link{weibull_distribution}}
}
\concept{distribution functions}
