% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_mean}
\alias{normal_mean}
\title{Normal Mean}
\usage{
normal_mean(mu)
}
\arguments{
\item{mu}{Numeric. Mean parameter.}
}
\value{
Numeric scalar, the mean of the distribution.
}
\description{
Returns the mean parameter of the Normal distribution.
}
\examples{
normal_mean(mu = 5)

}
