% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_median}
\alias{normal_median}
\title{Normal Median}
\usage{
normal_median(mu)
}
\arguments{
\item{mu}{Numeric. Mean parameter.}
}
\value{
Numeric scalar, the median of the distribution.
}
\description{
Returns the median of the Normal distribution.
}
\details{
For Normal distributions, the median equals the mean.
}
\examples{
normal_median(mu = 5)

}
