% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_mode}
\alias{normal_mode}
\title{Normal Mode}
\usage{
normal_mode(mu)
}
\arguments{
\item{mu}{Numeric. Mean parameter.}
}
\value{
Numeric scalar, the mode of the distribution.
}
\description{
Returns the mode of the Normal distribution.
}
\details{
For Normal distributions, the mode equals the mean.
}
\examples{
normal_mode(mu = 5)

}
