% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_pdf}
\alias{normal_pdf}
\title{Normal Probability Density Function}
\usage{
normal_pdf(x, mu, sigma)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{mu}{Numeric. Mean parameter.}

\item{sigma}{Numeric. Standard deviation parameter (must be positive).}
}
\value{
Numeric vector of density values.
}
\description{
Computes the PDF of the Normal distribution.
}
\details{
The PDF is given by:
\deqn{f(x; \mu, \sigma) = \frac{1}{\sigma\sqrt{2\pi}} \exp\left(-\frac{(x-\mu)^2}{2\sigma^2}\right)}
}
\examples{
normal_pdf(x = -3:3, mu = 0, sigma = 1)

}
