% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_normal.R
\name{normal_variance}
\alias{normal_variance}
\title{Normal Variance}
\usage{
normal_variance(sigma)
}
\arguments{
\item{sigma}{Numeric. Standard deviation parameter (must be positive).}
}
\value{
Numeric scalar, the variance of the distribution.
}
\description{
Computes the variance of the Normal distribution.
}
\details{
The variance is:
\deqn{\text{Var}(X) = \sigma^2}
}
\examples{
normal_variance(sigma = 2)

}
