% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_info.R, R/loglik_methods.R
\docType{class}
\name{parameter_info}
\alias{parameter_info}
\alias{parameter_info-class}
\title{Parameter Information Structure}
\usage{
parameter_info(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Object of class \code{parameter_info}.
}
\description{
S3 class defining the structure to hold model parameter details for likelihood-based fitting.
}
\details{
Objects of this class contain:
\itemize{
\item{\code{parameter_values}: Named list of initial parameter values used in modeling.}
\item{\code{lower_bounds}: Lower bounds for parameter estimates.}
\item{\code{upper_bounds}: Upper bounds for parameter estimates.}
\item{\code{estimate_flag}: Logical or numeric vector (1/0) indicating whether each
parameter is estimated (1) or fixed (0).}
\item{\code{link_functions}: List of link functions for each parameter.}
\item{\code{formulas}: List of formulas describing model structure for each parameter.}
\item{\code{design_matrices}: List of design matrices generated from formulas and data.}
\item{\code{model_data}: Data frame used to create design matrices.}
\item{\code{sep}: String defining separator for compound parameter names.}
}
}
\seealso{
\code{\link{create_parameter_info}}, \code{\link{summary.parameter_info}}
}
