% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{print.boundary_check}
\alias{print.boundary_check}
\title{Print method for boundary_check objects}
\usage{
\method{print}{boundary_check}(x, ...)
}
\arguments{
\item{x}{Result from check_boundary_parameters()}

\item{...}{Additional arguments (unused)}
}
\description{
Displays results from check_boundary_parameters() in a readable format.
}
