% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_star_center_points.R
\name{print.central_composite}
\alias{print.central_composite}
\title{Print method for central composite designs}
\usage{
\method{print}{central_composite}(x, ...)
}
\arguments{
\item{x}{Central composite design from augment_with_star_center_points()}

\item{...}{Additional arguments}
}
\description{
Print method for central composite designs
}
