% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{print.loglik}
\alias{print.loglik}
\title{Print method for loglik fitted models}
\usage{
\method{print}{loglik}(x, ...)
}
\arguments{
\item{x}{Object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for loglik fitted models
}
