% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_control.R
\name{print.loglik_control}
\alias{print.loglik_control}
\title{Print method for loglik_control}
\usage{
\method{print}{loglik_control}(x, ...)
}
\arguments{
\item{x}{Object of class \code{loglik_control}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for loglik_control
}
