% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_inference.R
\name{print.lrt}
\alias{print.lrt}
\title{Print Method for LRT Results}
\usage{
\method{print}{lrt}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{lrt}.}

\item{digits}{Number of digits for printing (default 4).}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for LRT Results
}
