% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_design.R
\name{print.parameter_design}
\alias{print.parameter_design}
\title{Print method for parameter_design objects}
\usage{
\method{print}{parameter_design}(x, ...)
}
\arguments{
\item{x}{Object of class parameter_design}

\item{...}{Additional arguments}
}
\description{
Print method for parameter_design objects
}
