% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{sandwich}
\alias{sandwich}
\title{Robust Sandwich Variance-Covariance Matrix}
\usage{
sandwich(object, ...)
}
\arguments{
\item{object}{Object of class "loglik".}
}
\value{
Sandwich variance-covariance matrix.
}
\description{
Computes the sandwich (robust) covariance estimator.
}
