% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_info.R
\name{summary.parameter_info}
\alias{summary.parameter_info}
\title{Summary for "parameter_info" object}
\usage{
\method{summary}{parameter_info}(object)
}
\arguments{
\item{object}{A parameter_info object as returned by \code{create_parameter_info()}.}
}
\value{
A list with info and parameter details for inspection or printing.
}
\description{
Prints formula representation and tabulated parameter values.
}
