% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{terms.loglik}
\alias{terms.loglik}
\title{Extract Model Terms}
\usage{
\method{terms}{loglik}(x, ...)
}
\arguments{
\item{x}{An object of class \code{loglik}.}

\item{...}{Additional arguments (unused).}
}
\value{
List of terms objects for each parameter.
}
\description{
Extract Model Terms
}
