% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{trace_history}
\alias{trace_history}
\title{Extract Optimization Trace}
\usage{
trace_history(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Matrix with columns: step, criterion, parameters, gradients, log-likelihood.
}
\description{
Returns iteration history from the optimization process.
}
