% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_sandwich.R
\name{vcov_sandwich.loglik}
\alias{vcov_sandwich.loglik}
\title{Robust Sandwich Variance-Covariance Matrix}
\usage{
vcov_sandwich.loglik(object)
}
\arguments{
\item{object}{Object of class "loglik".}
}
\value{
Sandwich variance-covariance matrix.
}
\description{
Computes the sandwich (robust) covariance estimator.
}
