% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_info_methods.R
\name{as.data.frame.parameter_info}
\alias{as.data.frame.parameter_info}
\title{Title}
\usage{
\method{as.data.frame}{parameter_info}(parameter_info, data = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
