% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_methods.R
\name{gradient_matrix}
\alias{gradient_matrix}
\title{Extract Gradient/Score Matrix}
\usage{
gradient_matrix(object)
}
\arguments{
\item{object}{An object of class \code{loglik}.}
}
\value{
Matrix (n_obs x n_params) of score contributions.
}
\description{
Returns the matrix of score contributions (gradient for each observation).
}
